refactor(HorizontalRule): update atom
