--- src/base/static/components/atoms/misc.scss
@@ -1,4 +1,0 @@
-.mapseed__hr {
-  margin-top: 5px;
-  margin-bottom: 5px;
-}
