+++ src/base/static/components/molecules/map-layer-group.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import { connect } from "react-redux";
 import classNames from "classnames";
 
-import { HorizontalRule } from "../atoms/misc";
+import { HorizontalRule } from "../atoms/layout";
 import { Header5 } from "../atoms/typography";
 
 import { mapConfigSelector } from "../../state/ducks/map-config";
