+++ src/base/static/components/molecules/map-legend-group.scss
@@ -1,7 +1,3 @@
-.map-legend-group__hr {
-  margin: 15px 0 15px 0;
-}
-
 .map-legend-group__title {
   margin-bottom: 10px;
 }
